Update README to build python before the rest of the system
Right now there seems to be some clash of the python version shipped with
macOS and the gobject-introspection package which leads to an error during
build. A workaround is to build python before the rest of the system which
makes the build process use the built python instead of the system one
which eliminates the error.