Go simple with LANG determination
AppleLocale is set based on the country set in the "region"
settings and territory based on the first language in the
"preferred languages" settings.  In general, this should be
enough to set LANG and other language-related
environment variables for Geany.

The current script does complicated things to determine LANG but the
result isn't correct in all cases. LANG isn't set correctly in cases
where country doesn't correspond to the territory language. For
instance, when country is set to "cs" and territory to "DK", we get
"cs_DK" but for this combination no locale is found because it's under
the "cs_CZ" directory and because of this the language falls back into
"en_US".

Instead of doing fancy stuff, let's just set LANG and all the related
environment variables to the contents of AppleLocale and add the "UTF-8"
extension. GTK will get the best locale language for us and if nothing
is found, it falls back to English. And VTE will always get a valid
LANG settings.