Launch Geany using exec() in the binary launcher
The previous method lauched Geany using NSTask so the result was 2
processes - one for launcher and one for Geany. For MacOS the main
process was the launcher and when some events were sent to it,
it just hanged because the launcher was waiting for Geany's process
to terminate.

By using exec() we have just one process and events are sent correctly
to Geany.