Load libgeany using dlopen() instead of using exec() to run Geany
macOS Catalina apparently doesn't like when we first run the launcher,
which sets environment variables, and from it run Geany using exec().
It confuses the dock and finder which think Geany isn't running
and when opening a file, they create a new instance of Geany.

Instead of using the binary as a launcher, we can just replace Geany's
main binary and call the libgeany's main_lib() by ourselves. However,
statically linking the library doesn't work as the environment variables
are set too late (the libraries are already loaded and upon loading
they read the old values). Therefore, we have to load libgeany dynamically
using dlopen().